requirements.txt
Flask
werkzeug
mysql-connector-python
python-dotenv
prometheus_client
flask_prometheus_metrics